* device-src/s3-device.c: Move some declaration to s3-device.h * device-src/s3-device.h: New file. * device-src/Makefile.am: Add s3-device.h.
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@5807 a8d146d6-cc15-0410-8900-af154a0219e0
